+++ src/encoder_motion_integration.cpp
@@ -44,7 +44,7 @@ bool encoderMotionUpdate() {
   
   for (int i = 0; i < 4; i++) {
     if (wj66IsStale(i)) {
-      logWarning("Encoder %d stale (%lu ms)", i, wj66GetAxisAge(i));
+      logWarning("Encoder %d stale (%lu ms)", i, (unsigned long)wj66GetAxisAge(i));
       faultLogEntry(FAULT_WARNING, FAULT_ENCODER_TIMEOUT, i, wj66GetAxisAge(i), "Encoder Stale");
       all_valid = false;
       continue;
@@ -71,13 +71,13 @@ bool encoderMotionUpdate() {
         position_errors[i].error_time_ms = now; 
         position_errors[i].error_count++;
         
-        logWarning("Axis %d error: %ld (limit %ld)", i, error, encoder_error_threshold);
+        logWarning("Axis %d error: %ld (limit %ld)", i, (long)error, (long)encoder_error_threshold);
         faultLogEntry(FAULT_WARNING, FAULT_ENCODER_SPIKE, i, error, "Axis deviation"); 
       }
     } else {
       if (position_errors[i].error_active) {
         uint32_t duration = now - position_errors[i].error_time_ms;
-        logInfo("Axis %d error cleared (%lu ms)", i, duration);
+        logInfo("Axis %d error cleared (%lu ms)", i, (unsigned long)duration);
         position_errors[i].error_active = false;
         position_errors[i].error_time_ms = 0; 
       }
@@ -116,10 +116,11 @@ void encoderMotionDiagnostics() {
   Serial.printf("Threshold: %.1f mm\n", encoder_error_threshold / 1000.0f);
   
   for (int i = 0; i < 4; i++) {
+    // FIX: Cast for printf
     Serial.printf("Axis %d: Err=%.1f mm | Max=%.1f mm | State=%s | Dur=%lu ms\n",
         i, position_errors[i].current_error / 1000.0f, 
         position_errors[i].max_error / 1000.0f,
         position_errors[i].error_active ? "[ERR]" : "[OK]",
-        encoderMotionGetErrorDuration(i));
+        (unsigned long)encoderMotionGetErrorDuration(i));
   }
 }
